Add buffer before getting its name in diagnostic list
Fixes #11

I'm not 100% sure this is the right thing to do, to load the buffersj
strictly. It seems to me it would be nicer to just store the filename
here and only load the buffer if the user selects it. This also seems to
display the full path to the current buffer sometimes :/

I'm sure someone with better vim scripting skills can do a better job at
this.